% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{logistica_estudio_caso_control}
\alias{logistica_estudio_caso_control}
\title{Logística de Estudio de Casos y Controles}
\usage{
logistica_estudio_caso_control(
  n_casos,
  n_controles,
  tasa_identificacion_casos,
  tasa_seleccion_controles,
  tasa_rechazo_casos,
  tasa_rechazo_controles,
  dias_laborables_mes
)
}
\arguments{
\item{n_casos}{Número de casos requeridos}

\item{n_controles}{Número de controles requeridos}

\item{tasa_identificacion_casos}{Tasa de identificación de casos por día}

\item{tasa_seleccion_controles}{Tasa de selección de controles por día}

\item{tasa_rechazo_casos}{Tasa de rechazo esperada para casos}

\item{tasa_rechazo_controles}{Tasa de rechazo esperada para controles}

\item{dias_laborables_mes}{Número de días laborables por mes}
}
\value{
Una lista con los cálculos logísticos del estudio
}
\description{
Logística de Estudio de Casos y Controles
}
